Allows for saving and loading of histograms into calibration program for quicker repeated loading.

+++ lib/event_handler.py
@@ -23,7 +23,7 @@ class event_handler:
     def sort_events(self, event_queue, particle_hit_list):
         if self.sort_type == 'event':
             self.sort_event_based(event_queue, particle_hit_list)
-        elif self.sort_type =='raw':
+        elif self.sort_type == 'raw':
             print("RAW output selected")
             self.raw_sorter(event_queue, particle_hit_list)
         elif self.sort_type == 'histo':
@@ -38,9 +38,8 @@ class event_handler:
         if self.calibrate:
             particle_hit_list = self.energy_calibration.calibrate_list(particle_hit_list)
         for particle_hit in particle_hit_list:
-            #self.total_count = self.total_count + 1
             if particle_hit['chan'] not in self.histo_data_dict.keys():
-                self.histo_data_dict.update({particle_hit['chan']: np.zeros(self.max_pulse_height)})
+                self.histo_data_dict.update({particle_hit['chan']: np.zeros(self.max_pulse_height, dtype=int)})
             if particle_hit['pulse_height'] < self.max_pulse_height:
                 self.histo_data_dict[particle_hit['chan']][particle_hit['pulse_height']] = self.histo_data_dict[particle_hit['chan']][particle_hit['pulse_height']] + 1
         return 0
